pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.85.0",
    "google-genai>=1.0.0",
    "openai>=2.28.0",
    "python-dotenv>=1.2.2",
]
